Make rake completion work on Linux
There is (again) a slight discrepancy between how `stat` works on OSX and on Linux.
To get a file modification date:
  - OSX: stat -f%m
  - Linux: stat --format=%y

This tries to run `stat -f%m .` and detects the return value. If this command
fails, meaning that the -f option is invalid, fallback to --format=%y

Yes, this means a tiny delay on each zsh startup, but I don't think caching
the $stat_cmd would make any difference.